ModLog: fix posting null attachments for deleted message logs
If attachments are not given to `upload_log`, an empty list is used. By
default, `zip_longest` uses `None` ass the fill value, so each message
was getting paired with a `None` (AKA null) attachment. The filed in the
DB is non-nullable so an empty list must be used instead.

Fixes #792